+++ Assets/OkapiKit/Scripts/Hypertag/HypertaggedObject.cs
@@ -2,7 +2,6 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 using NaughtyAttributes;
-using UnityEngine.Events;
 
 namespace OkapiKit
 {
@@ -125,6 +124,13 @@ namespace OkapiKit
         {
             List<GameObject> ret = new List<GameObject>();
 
+            FindGameObjectsWithHypertag(tags, ret);
+
+            return ret;
+        }
+
+        public static List<GameObject> FindGameObjectsWithHypertag(Hypertag[] tags, List<GameObject> ret)
+        {
             foreach (var t in tags)
             {
                 hypertaggedObjects.TryGetValue(t, out var list);
@@ -141,6 +147,13 @@ namespace OkapiKit
         {
             List<GameObject> ret = new List<GameObject>();
 
+            FindGameObjectsWithHypertag(tag, ret);
+
+            return ret;
+        }
+
+        public static List<GameObject> FindGameObjectsWithHypertag(Hypertag tag, List<GameObject> ret)
+        {
             hypertaggedObjects.TryGetValue(tag, out var list);
             if (list != null)
             {